feat(bootstrap.py): Add `insert_headers()` function. - Move `insert_headers()` from `setup.py` to `bootstrap.py`, and update   it with object oriented setup.
